requirements.txt
mysql-connector-python
pandas
numpy
tensorflow
scikit-learn
psutil
matplotlib
tensorflow.keras.layers